+++ test/c++/gfs/clef_handle_bug.cpp
@@ -23,7 +23,7 @@ using namespace triqs;
 triqs::clef::placeholder<0> iw_;
 triqs::clef::placeholder<1> inu_;
 
-int nw      = 3;
+size_t nw      = 3;
 double beta = 20;
 
 int main() {
